Screen, unscreen, and delete multiple comments at once http://zilla.livejournal.org/show_bug.cgi?id=296
git-svn-id: http://code.livejournal.org/svn/livejournal/trunk@1888 1dfba125-df10-0410-8a77-af1a8d6d074f

+++ cgi-bin/talklib.pl
@@ -282,7 +282,6 @@ sub update_commentalter {
     my $dbcm = $dbcs->{'dbh'};
     my $clustered = $u->{'clusterid'};
     my $userid = $u->{'userid'};
-    LJ::load_props($dbs, "log");
     my $prop = LJ::get_prop("log", "commentalter");
 
     $itemid = $itemid + 0;
@@ -295,32 +294,31 @@ sub update_commentalter {
 }
 
 sub screen_comment {
-    # NOTE: assumes that the comment is indeed not screened
-    # this should be checked before calling screen_comment()
+    my $dbs = shift;
+    my $dbcs = shift;
+    my $u = shift;
+    my $itemid = shift(@_) + 0;
 
-    my ($dbs, $dbcs, $u, $itemid, $talkid) = @_;
+    my $in = join (',', map { $_+0 } @_);
+    return unless $in;
+
     my $dbcm = $dbcs->{'dbh'};
-    my $dbcr = $dbcs->{'reader'};
     my $clustered = $u->{'clusterid'};
-    my $userid = $u->{'userid'};
-    LJ::load_props($dbs, "log");
+    my $userid = $u->{'userid'} + 0;
     my $prop = LJ::get_prop("log", "hasscreened");
 
-    my $hasscreened;
     if ($clustered) {
-        $hasscreened = $dbcr->selectrow_array("SELECT COUNT(*) FROM talk2 " .
-                                              "WHERE journalid=$userid AND nodeid=$itemid AND nodetype='L' AND state='S'");
-        $dbcm->do("UPDATE talk2 SET state='S' WHERE journalid=$userid AND jtalkid=$talkid");
-        $dbcm->do("UPDATE log2 SET replycount=replycount-1 WHERE journalid=$userid AND jitemid=$itemid");
-        $dbcm->do("REPLACE INTO logprop2 (journalid, jitemid, propid, value) VALUES ($userid, $itemid, $prop->{'id'}, '1')")
-            unless $hasscreened;
+        my $updated = $dbcm->do("UPDATE talk2 SET state='S' WHERE journalid=$userid AND jtalkid IN ($in)");
+        if ($updated > 0) {
+            $dbcm->do("UPDATE log2 SET replycount=replycount-$updated WHERE journalid=$userid AND jitemid=$itemid");
+            $dbcm->do("REPLACE INTO logprop2 (journalid, jitemid, propid, value) VALUES ($userid, $itemid, $prop->{'id'}, '1')");
+        }
     } else {
-        $hasscreened = $dbcr->selectrow_array("SELECT COUNT(*) FROM talk WHERE ".
-                                              "nodeid=$itemid AND nodetype='L' AND state='S'");
-        $dbcm->do("UPDATE talk SET state='S' WHERE talkid=$talkid");
-        $dbcm->do("UPDATE log SET replycount=replycount-1 WHERE itemid=$itemid");
-        $dbcm->do("INSERT INTO logprop (itemid, propid, value) VALUES ($itemid, $prop->{'id'}, '1')")
-            unless $hasscreened;
+        my $updated = $dbcm->do("UPDATE talk SET state='S' WHERE talkid IN ($in)");
+        if ($updated > 0) {
+            $dbcm->do("UPDATE log SET replycount=replycount-$updated WHERE itemid=$itemid");
+            $dbcm->do("REPLACE INTO logprop (itemid, propid, value) VALUES ($itemid, $prop->{'id'}, '1')");
+        }
     }
 
     LJ::Talk::update_commentalter($dbs, $dbcs, $u, $itemid);
@@ -328,40 +326,44 @@ sub screen_comment {
 }
 
 sub unscreen_comment {
-    # NOTE: assumes that the comment is indeed screened
-    # this should be checked before calling unscreen_comment()
+    my $dbs = shift;
+    my $dbcs = shift;
+    my $u = shift;
+    my $itemid = shift(@_) + 0;
 
-    my ($dbs, $dbcs, $u, $itemid, $talkid) = @_;
+    my $in = join (',', map { $_+0 } @_);
+    return unless $in;
+
     my $dbcm = $dbcs->{'dbh'};
     my $dbcr = $dbcs->{'reader'};
     my $clustered = $u->{'clusterid'};
-    my $userid = $u->{'userid'};
-    LJ::load_props($dbs, "log");
+    my $userid = $u->{'userid'} + 0;
     my $prop = LJ::get_prop("log", "hasscreened");
 
-    my $hasscreened;
     if ($clustered) {
-        $hasscreened = $dbcr->selectrow_array("SELECT COUNT(*) FROM talk2 " .
-                                              "WHERE journalid=$userid AND nodeid=$itemid AND nodetype='L' AND state='S'");
-        $dbcm->do("UPDATE talk2 SET state='A' WHERE journalid=$userid AND jtalkid=$talkid");
-        $dbcm->do("UPDATE log2 SET replycount=replycount+1 WHERE journalid=$userid AND jitemid=$itemid");
-        $dbcm->do("DELETE FROM logprop2 WHERE journalid=$userid AND jitemid=$itemid AND propid=$prop->{'id'}")
-            if $hasscreened == 1;
+        my $updated = $dbcm->do("UPDATE talk2 SET state='A' WHERE journalid=$userid AND jtalkid IN ($in) AND state='S'");
+        if ($updated > 0) {
+            $dbcm->do("UPDATE log2 SET replycount=replycount+$updated WHERE journalid=$userid AND jitemid=$itemid");
 
+            my $hasscreened = $dbcm->selectrow_array("SELECT COUNT(*) FROM talk2 " .
+                                                     "WHERE journalid=$userid AND nodeid=$itemid AND nodetype='L' AND state='S'");
+            $dbcm->do("DELETE FROM logprop2 WHERE journalid=$userid AND jitemid=$itemid AND propid=$prop->{'id'}")
+                unless $hasscreened;
+        }
     } else {
-        $hasscreened = $dbcr->selectrow_array("SELECT COUNT(*) FROM talk WHERE ".
-                                              "nodeid=$itemid AND nodetype='L' AND state='S'");
-        $dbcm->do("UPDATE talk SET state='A' WHERE talkid=$talkid");
-        $dbcm->do("UPDATE log SET replycount=replycount+1 WHERE itemid=$itemid");
-        $dbcm->do("DELETE FROM logprop WHERE itemid=$itemid AND propid=$prop->{'id'}")
-            if $hasscreened == 1;
+        my $updated = $dbcm->do("UPDATE talk SET state='A' WHERE talkid IN ($in) AND state='S'");
+        if ($updated > 0) {
+            $dbcm->do("UPDATE log SET replycount=replycount+$updated WHERE itemid=$itemid");
+            my $hasscreened = $dbcr->selectrow_array("SELECT COUNT(*) FROM talk WHERE ".
+                                                     "nodeid=$itemid AND nodetype='L' AND state='S'");
+            $dbcm->do("DELETE FROM logprop WHERE itemid=$itemid AND propid=$prop->{'id'}")
+                unless $hasscreened;
+        }
     }
 
     LJ::Talk::update_commentalter($dbs, $dbcs, $u, $itemid);
     return;
 }
 
-
-    
 
 1;
